% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in data from a csv file
## A note from the author:
### Please leave feedback where appropriate.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character vector of filename of a csv file to read}
}
\value{
this function returns a dataframe of the data read in from a csv file
}
\description{
\code{fars_read} returns a data.frame with data from the supplised csv file.
}
\details{
The function takes a character vector of a path of a filename to read into R.
If the filename does not exist the function will stop and an error will be thrown.
}
\note{
fars_read depends on read_csv and tbl_df from the readr and dplyr packages respectively
}
\examples{
\dontrun{
  fars_read("data/filename.csv")
}


}
