% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarse data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years to read in and summarise}
}
\value{
a data.frame of summarised data which is converted to a wide format
}
\description{
\code{fars_summarize_years} takes a list of years and reads them in using the \code{fars_read_years}
it then binds these dataframes together and summarises the data.
}
\details{
The function will take in a vector of years and read in the data using the fars_summarize_years function,
it then binds these rows together and groups by the year and MONTH column creating a count column: n.
The data is then converted from a long format to a wide format using the spread function in tidyr.
}
\note{
This functions depends on the bind_rows, group_by, summarize functions from dplyr and spread from tidyr
}
\examples{
\dontrun{
  fars_summarize_years(c(2013, 2014))
}


}
