% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename based on year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be in the name of the file.}
}
\value{
A character vector filename
}
\description{
\code{make_filename} will create a character vector filename with an assocaited year input
}
\details{
This takes a year input and creates a filename with this year.
If the year is does not pass \code{as.integer} it will have a value of NA, and the function
will throw an error after being passed to sprintf.
}
\note{
This package does not depends on any extensions
}
\examples{
\dontrun{
  make_filename(2013)
}


}
